Added functionality to Match object to allow negative relative vsini.
This is achieved by broadening the target spectrum instead of the
reference spectra. The limits on vsini are updated to (-5, 10) km/s.
The format of the output .h5 files are changed to store the unmodified
and broadened target spectra, error checking is implemented to check
the version of specmatch used to create the files.